requirements.txt
motor~=3.0.0
PyYAML~=6.0
wfastcgi~=3.0.0.0
fastapi~=0.78.0
Jinja2~=3.1.2
elasticsearch~=8.2.2
fastapi-jwt-auth~=0.5.0
python-jose~=3.3.0
passlib~=1.7.4
uvicorn[standard]
humanize~=4.1.0
openxmllib-py3~=2.0.0
pyzmq~=23.0.0
moviepy~=1.0.3
ipython ~=7.33.0
matplotlib-inline~=0.1.3
opencv-python~=4.5.5.64
confluent-kafka~=1.8.2
pdf2image~=1.16.0
PyMuPDF~=1.19.6
matplotlib~=3.5.2
ocrmypdf~=13.4.5
Cython~=0.29.30
python-multipart~=0.0.5
pdfminer.six
pdfplumber~=0.7.1
watchdog~=2.1.8
itsdangerous~=2.1.2
opencv-python~=4.5.5.64
Pillow~=9.1.1
awaits~=0.0.1
dependency-injector~=4.39.1
boto3~=1.24.30
mypy-boto3-s3~=1.24.0